+++ R/tm_file_viewer.R
@@ -1,12 +1,14 @@
 #' File Viewer Teal Module
 #'
-#' The file viewer module provides a tool to upload and view static files.
+#' The file viewer module provides a tool to view static files.
 #' Supported formats include text formats, \code{PDF}, \code{PNG}, \code{APNG},
 #' \code{JPEG}, \code{SVG}, \code{WEBP}, \code{GIF} and \code{BMP}.
 #'
 #' @inheritParams teal::module
 #' @inheritParams teal.devel::standard_layout
-#' @param input_path (`list`) A list of the input path to either specific files of accepted formats or a directory.
+#' @param input_path (`list`) `optional` A list of the input paths to either: specific files of accepted formats,
+#'   a directory or a URL. The paths can be specified as absolute paths or relative to the running
+#'   directory of the application. Will default to current working directory if not supplied.
 #'
 #' @export
 #'
@@ -27,7 +29,7 @@
 #' }
 #'
 tm_file_viewer <- function(label = "File Viewer Module",
-                           input_path) {
+                           input_path = list("Current Working Directory" = ".")) {
   valid_url <- function(url_input, timeout = 2) {
     con <- try(url(url_input), silent = TRUE)
     check <- suppressWarnings(try(open.connection(con, open = "rt", timeout = timeout), silent = TRUE)[1])
@@ -91,11 +93,9 @@ srv_viewer <- function(input, output, session, datasets, input_path) {
     out <- tryCatch({
       readLines(con = selected_path)
     },
-    error = function(cond) {
-      return(FALSE)
-    },
+    error = function(cond) FALSE,
     warning = function(cond) {
-      return(FALSE)
+     `if`(grepl("^incomplete final line found on", cond[[1]]), suppressWarnings(eval(cond[[2]])), FALSE)
     }
     )
   }
@@ -140,28 +140,26 @@ srv_viewer <- function(input, output, session, datasets, input_path) {
   }
 
   tree_list <- function(file_or_dir) {
-    nested_list <- lapply(file_or_dir, function(y) {
-      file_class <- suppressWarnings(file(y))
+    nested_list <- lapply(file_or_dir, function(path) {
+      file_class <- suppressWarnings(file(path))
       close(file_class)
-
       if (class(file_class)[[1]] != "url") {
-        isdir <- file.info(y)$isdir
+        isdir <- file.info(path)$isdir
         if (!isdir) {
-          structure(y, sticon = "file")
+          structure(path, ancestry = path, sticon = "file")
         } else {
-          files <- list.files(y, full.names = TRUE, include.dirs = TRUE)
-
+          files <- list.files(path, full.names = TRUE, include.dirs = TRUE)
           out <- lapply(files, function(x) tree_list(x))
           out <- unlist(out, recursive = F)
           if (!is_empty(files)) names(out) <- basename(files)
           out
         }
       } else {
-        structure(y, sticon = "file")
+        structure(path, ancestry = path, sticon = "file")
       }
-
     })
-    names(nested_list) <- file_or_dir
+    missing_labels <- if (is.null(names(nested_list))) seq_along(nested_list) else which(names(nested_list) == "")
+    names(nested_list)[missing_labels] <- file_or_dir[missing_labels]
     nested_list
   }
 
@@ -175,11 +173,24 @@ srv_viewer <- function(input, output, session, datasets, input_path) {
     handlerExpr = {
       if (!is_empty(shinyTree::get_selected(input$tree))) {
         obj <- shinyTree::get_selected(input$tree, format = "names")[[1]]
-        selected_path <- do.call("file.path", as.list(c(attr(obj, "ancestry"), obj[1])))
-        req(selected_path)
+        repo <- attr(obj, "ancestry")
+        repo_collapsed <- if (length(repo) > 1) paste0(repo, collapse = "/") else repo
+        is_not_named <- file.exists(file.path(c(repo_collapsed, obj[1])))[1]
+
+        if (is_not_named) {
+          selected_path <- do.call("file.path", as.list(c(repo, obj[1])))
+        } else {
+          if (is_empty(repo)) {
+            selected_path <- do.call("file.path", as.list(attr(input$tree[[obj[1]]], "ancestry")))
+          } else {
+            selected_path <- do.call("file.path", as.list(attr(input$tree[[repo]][[obj[1]]], "ancestry")))
+          }
+        }
 
         output$output <- renderUI({
-          validate(need(!isTRUE(file.info(selected_path)$isdir), "Please select a single file."))
+          validate(
+            need(!isTRUE(file.info(selected_path)$isdir) && !is_empty(selected_path), "Please select a single file.")
+          )
           display_file(selected_path)
         })
       }